Remove cargo-culted normalization from basic auth
We were rejecting decoded secret keys containing non-ASCII characters.
I'm not aware of any S3-compatible service deranged enough to use these
but the unnecessary defensive check (which I imagine was cargo-culted
from some tutorial---or vibe coded with Claude *2*!) did give me pause
when reviewing this code trying to figure out why it was necessary. We
do still need the TextDecoder() dance because RFC 7617 *does* specify
credentials are UTF-8 and without it we'd be double-encoding.